refactor!: Role of `Registry` replaced by `EventpOpsAdd`
In the past, we used `Registry` to represent types that could be used as
parameters for `register_into()`. It was an independent, additional
trait created solely for this purpose, which could easily confuse
new users.

Now, we have extracted part of the functionality from `EventpOps` to
create `EventpOpsAdd`, which replaces the role of `Registry`.
This approach is more natural and easier to understand.

Signed-off-by: Fuu <g.for.xyz@gmail.com>